+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Trunk
 
+* project: new sample
 * project: add travis
 * api: rename plugin module by name
 * project: initial release